requirements_dsgl.txt
python-docx>=0.8.11
PyPDF2>=3.0.0
ebooklib>=0.18
beautifulsoup4>=4.11.0
openai>=1.0.0